[query] refactor: move create/drop table method from Database to Catalog
It is more smooth for query to use: it does not need to fetch a db
first the create/drop table.

It is a natual way since Catalog does not need share meta with Database
any more.

- fix: #2326
