+++ query/src/interpreters/interpreter_table_create.rs
@@ -45,9 +45,8 @@ impl Interpreter for CreateTableInterpreter {
     }
 
     async fn execute(&self) -> Result<SendableDataBlockStream> {
-        let datasource = self.ctx.get_catalog();
-        let database = datasource.get_database(self.plan.db.as_str())?;
-        database.create_table(self.plan.clone())?;
+        let catalog = self.ctx.get_catalog();
+        catalog.create_table(self.plan.clone())?;
 
         Ok(Box::pin(DataBlockStream::create(
             self.plan.schema.clone(),
